Code refactor
Created different classes, one for R2Api, one for binary stuff and
another for the debugger.

r = R2Api(r2pipe.open('/bin/ls'))
r.binary.analyzeCalls()
for f in r.functions():
    print f.name
r.debugger.start()
